+++ modules/core/test/lib/fetch/fetch-file.spec.js
@@ -53,9 +53,8 @@ test('readFileSync#file (TEXT)', t => {
 
 test('fetchFile#dataUrl', async t => {
   const response = await fetchFile(DATA_URL);
-  const data = await response.arrayBuffer();
-  t.ok(data, 'readFileSync loaded data url');
-  // t.equals(data, 'important content!', 'fetchFile loaded data url');
+  const data = await response.text();
+  t.equals(data, 'important content!', 'fetchFile loaded data url');
   t.end();
 });