+++ modules/images/test/lib/deprecated/binary-image-api-deprecated.spec.js
@@ -2,39 +2,28 @@
 // Copyright (c) 2016-17 Karl Cheng, MIT license
 
 import test from 'tape-promise/tape';
-import {fetchFile} from '@loaders.gl/core';
 
 import {isBinaryImage, getBinaryImageMIMEType, getBinaryImageSize} from '@loaders.gl/images';
-
-const readFile = url => fetchFile(url).then(response => response.arrayBuffer());
-
-const IMAGES = {};
-const IMAGES_PROMISE = Promise.all([
-  readFile('@loaders.gl/images/test/data/img1-preview.png').then(data => (IMAGES.png = data)),
-  readFile('@loaders.gl/images/test/data/img1-preview.jpeg').then(data => (IMAGES.jpeg = data)),
-  readFile('@loaders.gl/images/test/data/img1-preview.gif').then(data => (IMAGES.gif = data)),
-  readFile('@loaders.gl/images/test/data/img1-preview.bmp').then(data => (IMAGES.bmp = data))
-  // readFile('@loaders.gl/images/test/data/img1-preview.tiff').then(data => IMAGES.tiff = data)
-]).then(() => IMAGES);
+import {loadImages} from '../category-api/binary-image-api.spec';
 
 test('isBinaryImage (deprecated)', async t => {
-  const images = await IMAGES_PROMISE;
+  const images = await loadImages();
   for (const imageType in images) {
     t.equals(isBinaryImage(images[imageType]), true, `isBinaryImage(${imageType})`);
   }
-  t.equals(isBinaryImage(images.png, 'image/png'), true, 'isBinaryImage(png, png)');
-  t.equals(isBinaryImage(images.png, 'image/jpeg'), false, 'isBinaryImage(png, jpeg)');
-  t.equals(isBinaryImage(images.jpeg, 'image/png'), false, 'isBinaryImage(jpeg, png)');
-  t.equals(isBinaryImage(images.jpeg, 'image/jpeg'), true, 'isBinaryImage(jpeg, jpeg)');
+  t.equals(isBinaryImage(images['image/png'], 'image/png'), true, 'isBinaryImage(png, png)');
+  t.equals(isBinaryImage(images['image/png'], 'image/jpeg'), false, 'isBinaryImage(png, jpeg)');
+  t.equals(isBinaryImage(images['image/jpeg'], 'image/png'), false, 'isBinaryImage(jpeg, png)');
+  t.equals(isBinaryImage(images['image/jpeg'], 'image/jpeg'), true, 'isBinaryImage(jpeg, jpeg)');
   t.end();
 });
 
 test('getBinaryImageMIMEType (deprecated)', async t => {
-  const images = await IMAGES_PROMISE;
+  const images = await loadImages();
   for (const imageType in images) {
     t.equals(
       getBinaryImageMIMEType(images[imageType]),
-      `image/${imageType}`,
+      imageType,
       `getBinaryImageMIMEType(${imageType})`
     );
   }
@@ -42,7 +31,7 @@ test('getBinaryImageMIMEType (deprecated)', async t => {
 });
 
 test('getBinaryImageSize (deprecated)', async t => {
-  const images = await IMAGES_PROMISE;
+  const images = await loadImages();
   for (const imageType in images) {
     const dimensions = getBinaryImageSize(images[imageType]);
     t.equals(dimensions.width, 480, `width, should work with ${imageType.toUpperCase()} files`);