+++ modules/polyfills/src/fetch-node/response.node.js
@@ -1,30 +1,47 @@
-/* global TextDecoder */
-import fs from 'fs'; // `fs` will be empty object in browsers (see package.json "browser" field).
+/* global TextDecoder, TextEncoder */
+import assert from '../utils/assert';
+import {decompressReadStream, concatenateReadStream} from './utils/stream-utils.node';
 import Headers from './headers.node';
-import {readFile, createReadStream} from './read-file.node';
 
-const isDataURL = url => url.startsWith('data:');
-const isRequestURL = url => url.startsWith('http:') || url.startsWith('https:');
+const isBoolean = x => typeof x === 'boolean';
+const isFunction = x => typeof x === 'function';
+const isObject = x => x !== null && typeof x === 'object';
+const isReadableNodeStream = x =>
+  isObject(x) && isFunction(x.read) && isFunction(x.pipe) && isBoolean(x.readable);
 
-// Under Node.js we return a mock "fetch response object"
-// so that apps can use the same API as in the browser.
-//
-// Note: This is intended to be a lightweight implementation and will have limitations.
-// Apps that require more complete fech emulation in Node
-// are encouraged to use dedicated fetch polyfill modules.
+/**
+ * Polyfill for Browser Response
+ *
+ * Under Node.js we return a mock "fetch response object"
+ * so that apps can use the same API as in the browser.
+ *
+ * Note: This is intended to be a "lightweight" implementation and will have limitations.
+ *
+ * See https://developer.mozilla.org/en-US/docs/Web/API/Response
+ */
+import {Readable} from 'stream';
 
-// See https://developer.mozilla.org/en-US/docs/Web/API/Response
-export default class NodeFetchResponse {
-  constructor(url, options) {
+export default class Response {
+  // TODO - handle ArrayBuffer, ArrayBufferView, Buffer
+  constructor(body, options = {}) {
+    const {headers, status = 200, statusText, url} = options;
+
     this._url = url;
-    this._ok = true; // TODO - handle errors and set ok/status
-    this._status = 200;
-    this.options = options;
+    this._ok = status === 200;
+    this._status = status; // TODO - handle errors and set status
+    this._statusText = statusText;
+    this._headers = new Headers(headers);
     this.bodyUsed = false;
-    this._headers = null;
 
-    // TODO - is this used?
-    this.httpResponse = null;
+    // Check for content-encoding and create a decompression stream
+    if (isReadableNodeStream(body)) {
+      this._body = decompressReadStream(body, headers);
+    } else {
+      if (typeof body === 'string') {
+        body = new TextEncoder().encode(body);
+      }
+      this._body = Readable.from([body]);
+    }
   }
 
   // Subset of Properties
@@ -37,27 +54,33 @@ export default class NodeFetchResponse {
     return this._status;
   }
 
+  get statusText() {
+    return this._statusText;
+  }
+
   get url() {
     return this._url;
   }
 
   get headers() {
-    this._headers = this._headers || this._getHeaders();
     return this._headers;
   }
 
   // Returns a readable stream to the "body" of the response (or file)
   get body() {
-    const {url, options} = this;
+    assert(!this.bodyUsed);
+    assert(isReadableNodeStream(this._body)); // Not implemented: conversion of ArrayBuffer etc to stream
     this.bodyUsed = true;
-    return createReadStream(url, options);
+    return this._body;
   }
 
   // Subset of Methods
 
   async arrayBuffer() {
-    this.bodyUsed = true;
-    const data = await readFile(this.url, this.options);
+    if (!isReadableNodeStream(this._body)) {
+      return this._body;
+    }
+    const data = await concatenateReadStream(this._body);
     return data;
   }
 
@@ -70,47 +93,5 @@ export default class NodeFetchResponse {
   async json() {
     const text = await this.text();
     return JSON.parse(text);
-  }
-
-  // PRIVATE
-
-  _getHeaders() {
-    // Under Node.js we return a mock "fetch response object"
-    // so that apps can use the same API as in the browser.
-    //
-    // Note: This is intended to be a lightweight implementation and will have limitations.
-    // Apps that require more complete fech emulation in Node
-    // are encouraged to use dedicated fetch polyfill modules.
-
-    const headers = {};
-
-    if (this.httpResponse) {
-      const httpHeaders = this.httpResponse.getHeaders();
-      for (const name in httpHeaders) {
-        const header = headers[name];
-        headers[name] = String(header);
-      }
-    } else {
-      const contentLength = this._getContentLength();
-      if (Number.isFinite(contentLength)) {
-        headers['Content-Length'] = contentLength;
-      }
-    }
-
-    return new Headers(headers);
-  }
-
-  _getContentLength() {
-    if (isRequestURL(this.url)) {
-      // Needs to be read from actual headers
-      return null;
-    } else if (isDataURL(this.url)) {
-      // TODO - remove media type etc
-      return this.url.length - 'data:'.length;
-    }
-    // File URL
-    // TODO - how to handle non-existing file, this presumably just throws
-    const stats = fs.statSync(this.url);
-    return stats.size;
   }
 }